# bpo-46823: Implement LOAD_FAST__LOAD_ATTR_INSTANCE_VALUE