@sweeneyde 2022-02-22
pyperformance (gcc with --enable-optimizations --with-lto):
```
Slower (11):
- scimark_sparse_mat_mult: 5.17 ms +- 0.26 ms -> 5.34 ms +- 0.26 ms: 1.03x slower
- unpickle_list: 5.00 us +- 0.09 us -> 5.14 us +- 0.11 us: 1.03x slower
- xml_etree_parse: 159 ms +- 6 ms -> 163 ms +- 6 ms: 1.03x slower
- regex_effbot: 2.98 ms +- 0.07 ms -> 3.06 ms +- 0.03 ms: 1.03x slower
- float: 81.0 ms +- 1.9 ms -> 82.9 ms +- 1.8 ms: 1.02x slower
- go: 149 ms +- 3 ms -> 151 ms +- 3 ms: 1.01x slower
- unpack_sequence: 47.7 ns +- 0.6 ns -> 48.2 ns +- 1.2 ns: 1.01x slower
- dulwich_log: 81.6 ms +- 1.5 ms -> 82.3 ms +- 1.6 ms: 1.01x slower
- unpickle_pure_python: 259 us +- 5 us -> 261 us +- 5 us: 1.01x slower
- python_startup: 9.10 ms +- 0.14 ms -> 9.16 ms +- 0.16 ms: 1.01x slower
- sympy_sum: 186 ms +- 2 ms -> 187 ms +- 2 ms: 1.01x slower

Faster (16):
- deltablue: 4.38 ms +- 0.13 ms -> 4.09 ms +- 0.14 ms: 1.07x faster
- raytrace: 335 ms +- 39 ms -> 314 ms +- 6 ms: 1.07x faster
- scimark_monte_carlo: 72.5 ms +- 2.4 ms -> 69.5 ms +- 0.9 ms: 1.04x faster
- unpickle: 16.1 us +- 1.3 us -> 15.4 us +- 1.2 us: 1.04x faster
- logging_silent: 105 ns +- 4 ns -> 101 ns +- 2 ns: 1.04x faster
- scimark_sor: 120 ms +- 1 ms -> 117 ms +- 1 ms: 1.03x faster
- chaos: 79.3 ms +- 2.5 ms -> 76.8 ms +- 2.8 ms: 1.03x faster
- 2to3: 281 ms +- 24 ms -> 272 ms +- 7 ms: 1.03x faster
- fannkuch: 412 ms +- 8 ms -> 400 ms +- 5 ms: 1.03x faster
- pickle_dict: 28.1 us +- 0.5 us -> 27.3 us +- 0.7 us: 1.03x faster
- pidigits: 188 ms +- 1 ms -> 184 ms +- 2 ms: 1.02x faster
- scimark_lu: 117 ms +- 2 ms -> 114 ms +- 2 ms: 1.02x faster
- nbody: 102 ms +- 4 ms -> 100 ms +- 4 ms: 1.02x faster
- chameleon: 7.23 ms +- 0.19 ms -> 7.13 ms +- 0.13 ms: 1.01x faster
- regex_dna: 212 ms +- 2 ms -> 209 ms +- 2 ms: 1.01x faster
- xml_etree_process: 56.6 ms +- 0.9 ms -> 56.2 ms +- 0.9 ms: 1.01x faster

Benchmark hidden because not significant (32): crypto_pyaes, django_template, hexiom, html5lib, json_dumps, json_loads, logging_format, logging_simple, mako, meteor_contest, nqueens, pathlib, pickle, pickle_list, pickle_pure_python, pyflate, python_startup_no_site, regex_compile, regex_v8, richards, scimark_fft, spectral_norm, sqlalchemy_declarative, sqlalchemy_imperative, sqlite_synth, sympy_expand, sympy_integrate, sympy_str, telco, tornado_http, xml_etree_iterparse, xml_etree_generate

Geometric mean: 1.01x faster
```