@sweeneyde 2022-02-22
With a very nice microbenchmark:

```
Mean +- std dev: [main_micro_load_attr2] 6.49 ns +- 0.08 ns -> [pr_micro_load_attr2] 4.67 ns +- 0.12 ns: 1.39x faster
```

```python
from pyperf import Runner, perf_counter
from itertools import repeat

class A:
    pass

def bench(loops):
    a = A()
    a.x = 42
    it = repeat(None, loops)
    t0 = perf_counter()
    for x in it:
        a.x; a.x; a.x; a.x; a.x
        a.x; a.x; a.x; a.x; a.x
        a.x; a.x; a.x; a.x; a.x
        a.x; a.x; a.x; a.x; a.x
        a.x; a.x; a.x; a.x; a.x
    return perf_counter() - t0

runner = Runner()
runner.bench_time_func("a.x", bench, inner_loops=25)
```